+++ packages/Webkul/Core/src/Core.php
@@ -28,7 +28,7 @@ class Core
      *
      * @var string
      */
-    const BAGISTO_VERSION = '2.5.0-beta4';
+    const BAGISTO_VERSION = '2.5.0-beta5';
 
     /**
      * Current Channel.